refactor(monitoring): extract status data models from status_reporter
status_reporter.py (1208 lines) mixed ~20 pure dataclass data contracts
(BotStatus and its nested status models) with the StatusReporter behavior that
populates and writes them. Split the contracts (plus their DecimalEncoder and
timestamp/clock helpers) into a focused status_models.py, leaving
status_reporter.py to own behavior only.

status_reporter re-exports the models, so all 30 import sites are unchanged and
the model classes keep identical identity (verified). Two tests that froze time
by patching the module-global get_clock now patch it where the helper lives
(status_models). No behavior change.
